requirements.txt
django==1.11.1
djangorestframework==3.7.1
